Add validation for org tokens
Currently the validate function on the admin struct checks for email,
first_name, and last_name. This works for user tokens, but will fail
with an org token which /api/v1/self responds with like the following
```
{
    "name": "test-token",
    "privileges": [
        {
            "scope": "org",
            "org_id": "redacted",
            "role": "admin",
            "name": "Test Org"
        }
    ]
}
```

This commit adds a name field to the admin struct and also allows the
validate function to skip validation for email, first_name, and
last_name if the name field is present.